# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/cancel_collect_feed": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "取消收藏动态",
                "parameters": [
                    {
                        "description": "动态ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RemoveFeedFromCollectionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/user/cancel_collect_wallpaper": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "取消收藏壁纸",
                "parameters": [
                    {
                        "description": "壁纸数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RemoveWallpaperFromCollectionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/user/collect_feed": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "收藏动态",
                "parameters": [
                    {
                        "description": "动态ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddFeedToCollectionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/user/collect_wallpaper": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "收藏壁纸",
                "parameters": [
                    {
                        "description": "壁纸数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddWallpaperToCollectionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
...
        "/user/follow_user": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "关注用户",
                "parameters": [
                    {
                        "description": "用户ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.FollowUserPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
...
        "/user/profile": {
            "get": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "获取本人用户信息",
                "responses": {
                    "200": {
                        "description": "ok\" \"本人用户信息",
                        "schema": {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    }
                }
            },
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "更新本人用户信息",
                "parameters": [
                    {
                        "description": "昵称和头像",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateProfilePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"更新后的本人用户信息",
                        "schema": {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    }
                }
            }
        },
        "/user/sign_in": {
            "post": {
                "description": "微信授权登录｜Github授权登录",
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.LoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"Token和用户信息",
                        "schema": {
                            "$ref": "#/definitions/entity.LoginRes"
                        }
                    }
                }
            }
        },
        "/user/unfollow_user": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "取关用户",
                "parameters": [
                    {
                        "description": "用户ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UnfollowUserPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/user/user_info": {
            "get": {
                "description": "使用但不校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "获取其他人用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"用户信息",
                        "schema": {
                            "$ref": "#/definitions/entity.UserItem"
                        }
                    }
                }
            }
        },
        "/user/yd_callback": {
            "post": {
                "description": "提供给易登侧调用，获取用户扫码状态",
                "produces": [
                    "application/json"
                ],
                "summary": "易登回调",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.YDCallbackPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/user/yd_login_status": {
            "get": {
                "description": "用于微信扫码登录",
                "produces": [
                    "application/json"
                ],
                "summary": "检查易登状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "易登临时ID",
                        "name": "temp_user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"是否成功登录",
                        "schema": {
                            "$ref": "#/definitions/entity.GetYDLoginStatusResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.AddFeedToCollectionPayload": {
            "type": "object",
            "required": [
                "feed_id"
            ],
            "properties": {
                "feed_id": {
                    "type": "string"
                }
            }
        },
        "entity.AddWallpaperToCollectionPayload": {
            "type": "object",
            "required": [
                "wallpaper"
            ],
            "properties": {
                "wallpaper": {
                    "$ref": "#/definitions/model.Wallpaper"
                }
            }
        },
...
        "entity.FollowUserPayload": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.GetYDLoginStatusResp": {
            "type": "object",
            "properties": {
                "has_login": {
                    "type": "boolean"
                }
            }
        },
...
        "entity.LoginPayload": {
            "type": "object",
            "required": [
                "code",
                "type"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "entity.LoginRes": {
            "type": "object",
            "properties": {
                "associated": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "collected_feed_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "collected_wallpapers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Wallpaper"
                    }
                },
                "creation_time": {
                    "type": "string"
                },
                "followed_user_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "github_id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wechat_id": {
                    "type": "string"
                }
            }
        },
        "entity.Profile": {
            "type": "object",
            "properties": {
                "associated": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "github_id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wechat_id": {
                    "type": "string"
                }
            }
        },
        "entity.RemoveFeedFromCollectionPayload": {
            "type": "object",
            "required": [
                "feed_id"
            ],
            "properties": {
                "feed_id": {
                    "type": "string"
                }
            }
        },
        "entity.RemoveWallpaperFromCollectionPayload": {
            "type": "object",
            "required": [
                "wallpaper_id"
            ],
            "properties": {
                "wallpaper_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.UnfollowUserPayload": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.UpdateProfilePayload": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "entity.UserItem": {
            "type": "object",
            "properties": {
                "associated": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "collected": {
                    "type": "boolean"
                },
                "creation_time": {
                    "type": "string"
                },
                "github_id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wechat_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.WeChatUserInfo": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "openId": {
                    "type": "string"
                }
            }
        },
        "entity.YDCallbackPayload": {
            "type": "object",
            "required": [
                "tempUserId"
            ],
            "properties": {
                "cancelLogin": {
                    "type": "boolean"
                },
                "scanSuccess": {
                    "type": "boolean"
                },
                "tempUserId": {
                    "type": "string"
                },
                "wxMaUserInfo": {
                    "$ref": "#/definitions/entity.WeChatUserInfo"
                }
            }
        },
...
        "model.Tag": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "purity": {
                    "type": "string"
                }
            }
        },
        "model.Thumbs": {
            "type": "object",
            "properties": {
                "large": {
                    "type": "string"
                },
                "original": {
                    "type": "string"
                },
                "small": {
                    "type": "string"
                }
            }
        },
        "model.Wallpaper": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "colors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "dimension_x": {
                    "type": "integer"
                },
                "dimension_y": {
                    "type": "integer"
                },
                "favorites": {
                    "type": "integer"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "purity": {
                    "type": "string"
                },
                "ratio": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "thumbs": {
                    "$ref": "#/definitions/model.Thumbs"
                },
                "url": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        }
    }
}